requirements.txt
attrs==17.4.0
cycler==0.10.0
decorator==4.2.1
imutils==0.4.5
Jinja2==2.10
MarkupSafe==1.0
matplotlib==2.1.2
networkx==2.1
numpy==1.14.1
opencv-contrib-python==3.4.0.12
pandas==0.22.0
Pillow==5.0.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.3
PyWavelets==0.5.2
scikit-image==0.13.1
scipy==1.0.0
selenium==3.9.0
six==1.11.0
